+++ response/unprocessableEntity.go
@@ -1,6 +1,6 @@
 package response
 
-func UnprocessableEntity(message string) (status int16, res commonResponse) {
+func UnprocessableEntity(message string) (status int, res commonResponse) {
 	meg := "Unprocessable Entity"
 	if message != "" {
 		meg = message
